+++ libnopegl/src/api.c
@@ -344,7 +344,6 @@ int ngli_ctx_set_scene(struct ngl_ctx *s, struct ngl_scene *scene)
     ngpu_ctx_wait_idle(s->gpu_ctx);
     reset_scene(s, NGLI_ACTION_UNREF_SCENE);
 
-    s->default_graphics_state = NGPU_GRAPHICS_STATE_DEFAULTS;
     s->default_rendertarget_layout = *ngpu_ctx_get_default_rendertarget_layout(s->gpu_ctx);
 
     int ret = ngpu_ctx_begin_update(s->gpu_ctx);