+++ libnopegl/src/node_drawcolor.c
@@ -24,9 +24,9 @@
 #include <stddef.h>
 #include <string.h>
 
-#include "blend_mode.h"
 #include "filterschain.h"
 #include "geometry.h"
+#include "graphics_state.h"
 #include "image.h"
 #include "internal.h"
 #include "log.h"
@@ -87,7 +87,7 @@ struct drawcolor_opts {
     float color[3];
     struct ngl_node *opacity_node;
     float opacity;
-    enum ngli_blend_mode blend_mode;
+    struct ngli_graphics_state_opts state;
     struct ngl_node *geometry;
     struct ngl_node **filters;
     size_t nb_filters;
@@ -145,9 +145,7 @@ static const struct node_param drawcolor_params[] = {
     {"opacity",  NGLI_PARAM_TYPE_F32, OFFSET(opacity_node), {.f32=1.f},
                  .flags=NGLI_PARAM_FLAG_ALLOW_LIVE_CHANGE | NGLI_PARAM_FLAG_ALLOW_NODE,
                  .desc=NGLI_DOCSTRING("opacity of the color")},
-    {"blend_mode", NGLI_PARAM_TYPE_SELECT, OFFSET(blend_mode),
-                 .choices=&ngli_blend_mode_choices,
-                 .desc=NGLI_DOCSTRING("define how this node is composited with the current framebuffer")},
+    NGLI_GRAPHICS_STATE_PARAMS(state),
     {"geometry", NGLI_PARAM_TYPE_NODE, OFFSET(geometry),
                  .node_types=GEOMETRY_TYPES_LIST,
                  .desc=NGLI_DOCSTRING("geometry to be rasterized")},
@@ -261,7 +259,6 @@ static int drawcolor_init(struct ngl_node *node)
 }
 
 static int drawcolor_prepare(struct ngl_node *node,
-                             const struct ngpu_graphics_state *graphics_state,
                              const struct ngpu_rendertarget_layout *rendertarget_layout)
 {
     struct ngl_ctx *ctx = node->ctx;
@@ -325,9 +322,9 @@ static int drawcolor_prepare(struct ngl_node *node,
     s->vert_block_index = ngpu_pgcraft_get_block_index(s->crafter, "vert_params", NGPU_PROGRAM_STAGE_VERT);
     s->frag_block_index = ngpu_pgcraft_get_block_index(s->crafter, "frag_params", NGPU_PROGRAM_STAGE_FRAG);
 
-    /* Apply blend mode */
-    struct ngpu_graphics_state state = *graphics_state;
-    ret = ngli_blend_mode_apply(&state, o->blend_mode);
+    /* Apply the graphics state options */
+    struct ngpu_graphics_state state;
+    ret = ngli_graphics_state_init_from_opts(gpu_ctx, &state, &o->state);
     if (ret < 0)
         return ret;
 
@@ -478,12 +475,19 @@ static void drawcolor_uninit(struct ngl_node *node)
         ngli_geometry_freep(&s->geometry);
 }
 
+static uint32_t drawcolor_get_renderpass_usage(const struct ngl_node *node)
+{
+    const struct drawcolor_opts *o = node->opts;
+    return ngli_graphics_state_get_renderpass_usage(&o->state);
+}
+
 const struct node_class ngli_drawcolor_class = {
     .id        = NGL_NODE_DRAWCOLOR,
     .category  = NGLI_NODE_CATEGORY_DRAW,
     .name      = "DrawColor",
     .init      = drawcolor_init,
     .prepare   = drawcolor_prepare,
+    .get_renderpass_usage = drawcolor_get_renderpass_usage,
     .update    = ngli_node_update_children,
     .draw      = drawcolor_draw,
     .uninit    = drawcolor_uninit,