+++ libnopegl/src/node_drawmask.c
@@ -24,9 +24,9 @@
 #include <stddef.h>
 #include <string.h>
 
-#include "blend_mode.h"
 #include "filterschain.h"
 #include "geometry.h"
+#include "graphics_state.h"
 #include "internal.h"
 #include "log.h"
 #include <ngpu/ngpu.h>
@@ -85,7 +85,7 @@ struct drawmask_opts {
     struct ngl_node *content;
     struct ngl_node *mask;
     int inverted;
-    enum ngli_blend_mode blend_mode;
+    struct ngli_graphics_state_opts state;
     struct ngl_node *geometry;
     struct ngl_node **filters;
     size_t nb_filters;
@@ -146,9 +146,7 @@ static const struct node_param drawmask_params[] = {
                  .desc=NGLI_DOCSTRING("texture serving as mask (only the red channel is used)")},
     {"inverted",  NGLI_PARAM_TYPE_BOOL, OFFSET(inverted),
                  .desc=NGLI_DOCSTRING("whether to dig into or keep")},
-    {"blend_mode", NGLI_PARAM_TYPE_SELECT, OFFSET(blend_mode),
-                 .choices=&ngli_blend_mode_choices,
-                 .desc=NGLI_DOCSTRING("define how this node is composited with the current framebuffer")},
+    NGLI_GRAPHICS_STATE_PARAMS(state),
     {"geometry", NGLI_PARAM_TYPE_NODE, OFFSET(geometry),
                  .node_types=GEOMETRY_TYPES_LIST,
                  .desc=NGLI_DOCSTRING("geometry to be rasterized")},
@@ -273,7 +271,6 @@ static int drawmask_init(struct ngl_node *node)
 }
 
 static int drawmask_prepare(struct ngl_node *node,
-                            const struct ngpu_graphics_state *graphics_state,
                             const struct ngpu_rendertarget_layout *rendertarget_layout)
 {
     struct ngl_ctx *ctx = node->ctx;
@@ -364,9 +361,9 @@ static int drawmask_prepare(struct ngl_node *node,
     s->vert_block_index = ngpu_pgcraft_get_block_index(s->crafter, "vert_params", NGPU_PROGRAM_STAGE_VERT);
     s->frag_block_index = ngpu_pgcraft_get_block_index(s->crafter, "frag_params", NGPU_PROGRAM_STAGE_FRAG);
 
-    /* Apply blend mode */
-    struct ngpu_graphics_state state = *graphics_state;
-    ret = ngli_blend_mode_apply(&state, o->blend_mode);
+    /* Apply the graphics state options */
+    struct ngpu_graphics_state state;
+    ret = ngli_graphics_state_init_from_opts(gpu_ctx, &state, &o->state);
     if (ret < 0)
         return ret;
 
@@ -523,12 +520,19 @@ static void drawmask_uninit(struct ngl_node *node)
         ngli_geometry_freep(&s->geometry);
 }
 
+static uint32_t drawmask_get_renderpass_usage(const struct ngl_node *node)
+{
+    const struct drawmask_opts *o = node->opts;
+    return ngli_graphics_state_get_renderpass_usage(&o->state);
+}
+
 const struct node_class ngli_drawmask_class = {
     .id        = NGL_NODE_DRAWMASK,
     .category  = NGLI_NODE_CATEGORY_DRAW,
     .name      = "DrawMask",
     .init      = drawmask_init,
     .prepare   = drawmask_prepare,
+    .get_renderpass_usage = drawmask_get_renderpass_usage,
     .update    = ngli_node_update_children,
     .draw      = drawmask_draw,
     .uninit    = drawmask_uninit,