+++ libnopegl/src/node_drawtexture.c
@@ -24,9 +24,9 @@
 #include <stddef.h>
 #include <string.h>
 
-#include "blend_mode.h"
 #include "filterschain.h"
 #include "geometry.h"
+#include "graphics_state.h"
 #include "internal.h"
 #include "log.h"
 #include <ngpu/ngpu.h>
@@ -97,7 +97,7 @@ const struct param_choices texture_wrap_choices = {
 struct drawtexture_opts {
     struct ngl_node *texture_node;
     int wrap;
-    enum ngli_blend_mode blend_mode;
+    struct ngli_graphics_state_opts state;
     struct ngl_node *geometry;
     struct ngl_node **filters;
     size_t nb_filters;
@@ -154,9 +154,7 @@ static const struct node_param drawtexture_params[] = {
     {"wrap",     NGLI_PARAM_TYPE_SELECT, OFFSET(wrap),
                  .choices=&texture_wrap_choices,
                  .desc=NGLI_DOCSTRING("texture wrap behaviour")},
-    {"blend_mode", NGLI_PARAM_TYPE_SELECT, OFFSET(blend_mode),
-                 .choices=&ngli_blend_mode_choices,
-                 .desc=NGLI_DOCSTRING("define how this node is composited with the current framebuffer")},
+    NGLI_GRAPHICS_STATE_PARAMS(state),
     {"geometry", NGLI_PARAM_TYPE_NODE, OFFSET(geometry),
                  .node_types=GEOMETRY_TYPES_LIST,
                  .desc=NGLI_DOCSTRING("geometry to be rasterized")},
@@ -277,7 +275,6 @@ static int drawtexture_init(struct ngl_node *node)
 }
 
 static int drawtexture_prepare(struct ngl_node *node,
-                             const struct ngpu_graphics_state *graphics_state,
                              const struct ngpu_rendertarget_layout *rendertarget_layout)
 {
     struct ngl_ctx *ctx = node->ctx;
@@ -359,8 +356,8 @@ static int drawtexture_prepare(struct ngl_node *node,
     s->vert_block_index = ngpu_pgcraft_get_block_index(s->crafter, "vert_params", NGPU_PROGRAM_STAGE_VERT);
     s->frag_block_index = ngpu_pgcraft_get_block_index(s->crafter, "frag_params", NGPU_PROGRAM_STAGE_FRAG);
 
-    struct ngpu_graphics_state state = *graphics_state;
-    ret = ngli_blend_mode_apply(&state, o->blend_mode);
+    struct ngpu_graphics_state state;
+    ret = ngli_graphics_state_init_from_opts(gpu_ctx, &state, &o->state);
     if (ret < 0)
         return ret;
 
@@ -514,12 +511,19 @@ static void drawtexture_uninit(struct ngl_node *node)
         ngli_geometry_freep(&s->geometry);
 }
 
+static uint32_t drawtexture_get_renderpass_usage(const struct ngl_node *node)
+{
+    const struct drawtexture_opts *o = node->opts;
+    return ngli_graphics_state_get_renderpass_usage(&o->state);
+}
+
 const struct node_class ngli_drawtexture_class = {
     .id        = NGL_NODE_DRAWTEXTURE,
     .category  = NGLI_NODE_CATEGORY_DRAW,
     .name      = "DrawTexture",
     .init      = drawtexture_init,
     .prepare   = drawtexture_prepare,
+    .get_renderpass_usage = drawtexture_get_renderpass_usage,
     .update    = ngli_node_update_children,
     .draw      = drawtexture_draw,
     .uninit    = drawtexture_uninit,